nopbai

+++ src/main/java/com/codegym/controller/StaffRestController.java
@@ -40,7 +40,7 @@ public class StaffRestController {
                                          @RequestParam(defaultValue = "") String name) {
         System.out.println("a");
         Pageable pageable = PageRequest.of(page, 5);
-        Page<Account> listStaff = iAccountService.findAllPage(pageable,name);
+        Page<Account> listStaff = iAccountService.findAllPage(pageable, name);
         return new ResponseEntity<>(listStaff, HttpStatus.OK);
     }
 
@@ -52,14 +52,14 @@ public class StaffRestController {
         String result = "";
         //1 tuong ung trung code, 2 trung username, 3 trung can cuoc
         for (int i = 0; i < accountList.size(); i++) {
-            if (account.getCode().equals(accountList.get(i).getCode())){
+            if (account.getCode().equals(accountList.get(i).getCode())) {
                 result += "1";
             }
-            if (account.getUsername().equals(accountList.get(i).getUsername())){
+            if (account.getUsername().equals(accountList.get(i).getUsername())) {
                 result += "2";
             }
-            if (account.getIdentityCode().equals(accountList.get(i).getIdentityCode())){
-                result+= "3";
+            if (account.getIdentityCode().equals(accountList.get(i).getIdentityCode())) {
+                result += "3";
             }
         }
 
@@ -84,7 +84,7 @@ public class StaffRestController {
     public ResponseEntity<?> updateStaff(@PathVariable Integer id, @RequestBody Account account) {
         List<Account> accountList = iAccountService.findAll();
         for (int i = 0; i < accountList.size(); i++) {
-            if (accountList.get(i).getId().equals(id)){
+            if (accountList.get(i).getId().equals(id)) {
                 accountList.remove(i);
                 break;
             }
@@ -92,14 +92,14 @@ public class StaffRestController {
         String result = "";
         //1 tuong ung trung code, 2 trung username, 3 trung can cuoc
         for (int i = 0; i < accountList.size(); i++) {
-            if (account.getCode().equals(accountList.get(i).getCode())){
-                result += "1";
-            }
-            if (account.getUsername().equals(accountList.get(i).getUsername())){
-                result += "2";
-            }
-            if (account.getIdentityCode().equals(accountList.get(i).getIdentityCode())){
-                result+= "3";
+//            if (account.getCode().equals(accountList.get(i).getCode())){
+//                result += "1";
+//            }
+//            if (account.getUsername().equals(accountList.get(i).getUsername())){
+//                result += "2";
+//            }
+            if (account.getIdentityCode().equals(accountList.get(i).getIdentityCode())) {
+                result += "3";
             }
         }
 
@@ -107,7 +107,7 @@ public class StaffRestController {
             String enc = passwordEncoder.encode(account.getPassword());
             account.setPassword(enc);
             iAccountService.save(account);
-            iAccountService.save(account);
+//            iAccountService.save(account);
             return new ResponseEntity<>(HttpStatus.CREATED);
         }
 
@@ -116,26 +116,45 @@ public class StaffRestController {
 
     @DeleteMapping("/remove/{id}")
     public ResponseEntity<?> removeStaff(@PathVariable Integer id) {
-        Account account = iAccountService.findById(id);
-        iAccountService.delete(account);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-    @GetMapping("cote-a")
-    public ResponseEntity<?> findCote(){
-        List<Cote> coteList = iAccountService.findCoteManagement("trucvi");
-        return new ResponseEntity<>(coteList,HttpStatus.OK);
+        boolean check = true;
+        List<Cote> coteList = iAccountService.findAllCote();
+        for (int i = 0; i < coteList.size(); i++) {
+            if (coteList.get(i).getAccount().getId() == id) {
+                check = true;
+                break;
+            } else {
+                check = false;
+
+            }
+
+        }
+        if(!check){
+            Account account = iAccountService.findById(id);
+            iAccountService.delete(account);
+        }
+
+//        Account account = iAccountService.findById(id);
+//        iAccountService.delete(account);
+        return new ResponseEntity<>(check,HttpStatus.OK);
     }
 
+//    @GetMapping("cote-a")
+//    public ResponseEntity<?> findCote(@RequestBody Integer id) {
+//        List<Cote> coteList = iAccountService.findCoteManagement(id);
+//        return new ResponseEntity<>(coteList, HttpStatus.OK);
+//    }
+    @GetMapping("cote-all")
+    public ResponseEntity<?> findAllCote() {
+        List<Cote> coteList = iAccountService.findAllCote();
+        return new ResponseEntity<>(coteList, HttpStatus.OK);
+    }
 
     @GetMapping("/detail/{id}")
-    public ResponseEntity<?> findDetail(@PathVariable Integer id){
+    public ResponseEntity<?> findDetail(@PathVariable Integer id) {
         Account account = iAccountService.findById(id);
         iAccountService.findByDetail(account);
-        return new ResponseEntity<>(account,HttpStatus.OK);
+        return new ResponseEntity<>(account, HttpStatus.OK);
     }
-
-
-
 
 
 }
